+++ sipx/backends/__init__.py
@@ -1,4 +1,15 @@
 from sipx.backends.mock import MockBackend
-from sipx.backends.native import NativeSipBackend
+from sipx.backends.native import (
+    NativeSipBackend,
+    NativeSipCall,
+    NativeSipCallError,
+    NativeSipCallState,
+)
 
-__all__ = ["MockBackend", "NativeSipBackend"]
+__all__ = [
+    "MockBackend",
+    "NativeSipBackend",
+    "NativeSipCall",
+    "NativeSipCallError",
+    "NativeSipCallState",
+]